pyproject.toml
[project]
name = "inno-map-access"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "geopandas>=1.1.3",
    "geoparquet-io>=0.4.0",
    "geozarr-toolkit>=0.1.2",
    "inno-data-access>=0.0.2",
    "matplotlib>=3.10.8",
    "numcodecs>=0.16.5",
    "numpy>=2.4.3",
    "obstore>=0.9.2",
    "odc-stac>=0.5.2",
    "omnicloudmask>=1.7.1",
    "planetary-computer>=1.0.0",
    "pystac-client>=0.9.0",
    "rasterio>=1.5.0",
    "requests>=2.32.5",
    "rioxarray>=0.22.0",
    "scipy>=1.17.1",
    "tqdm>=4.67.3",
    "unidecode>=1.4.0",
    "xbatcher>=0.4.0",
    "zarr[gpu,remote]>=3.1.6",
    "zarrs>=0.2.3",
]

[project.scripts]
inno-map-access = "inno_map_access:main"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"

[[tool.uv.index]]
url = "https://nexusrepo.natuurenbos.be/repository/anb-python-group/simple/"
name = "nexus"
authenticate = "always"
priority = "supplemental"


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4
